+++ instrumentation/gruf/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-instrumentation-gruf
 
+### v0.2.0 / 2024-02-20
+
+* ADDED: Add support gruf 2.19
+
 ### v0.1.1 / 2023-10-16
 
 * FIXED: Remove activesupport dependency